[proj-104] refactor: mapper를 utils에 이동
