+++ src/main/java/com/omnm/hanasset/loan/utils/LoanMapper.java
@@ -1,4 +1,4 @@
-package com.omnm.hanasset.loan;
+package com.omnm.hanasset.loan.utils;
 
 import com.omnm.hanasset.loan.dto.LoanInfoDTO;
 import com.omnm.hanasset.loan.entity.Loan;
